requirements.txt
Django==4.2.11
djangorestframework==3.15.1
django-phonenumber-field[phonenumbers]==4.11.0
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2